% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_pal.R
\name{cb_pal}
\alias{cb_pal}
\title{Get CB color palette}
\usage{
cb_pal(name = "cb_okabeito")
}
\arguments{
\item{name}{Name of the color palette}
}
\value{
A CB color palette
}
\description{
Return a CB color palette
}
