% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_palettes.R
\docType{data}
\name{cb_palettes}
\alias{cb_palettes}
\title{Full list of palettes}
\format{
An object of class \code{list} of length 12.
}
\usage{
cb_palettes
}
\description{
\code{names(cb_palettes)} returns the names of all palettes.
}
\keyword{datasets}
