% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_cb_base.R
\name{theme_cb_base}
\alias{theme_cb_base}
\title{CB base theme}
\usage{
theme_cb_base(
  base_family = "Atkinson Hyperlegible",
  base_size = 14,
  hjust = 0,
  text_clr_1 = "grey20",
  text_clr_2 = "grey40",
  text_clr_3 = "grey70",
  paper = "#FDFDFD"
)
}
\arguments{
\item{base_family}{Base font family. Default "Atkinson Hyperlegible".}

\item{base_size}{Base font size for text elements. Default 14.}

\item{hjust}{Horizontal alignment of title, subtitle and caption. Default 0.}

\item{text_clr_1}{Text color (1st). Default "grey20".}

\item{text_clr_2}{Text color (2nd). Default "grey40".}

\item{text_clr_3}{Text color (3rd). Default "grey80".}

\item{paper}{Background color. Default "white".}
}
\value{
A ggplot2 theme
}
\description{
Custom ggplot theme
}
